+++ src/customHooks/useOutsideAlerter.tsx
@@ -1,0 +1,20 @@
+import { RefObject, useEffect } from 'react';
+
+export function useOutsideAlerter(
+  ref: RefObject<HTMLElement>,
+  onClose: () => void,
+) {
+  useEffect(() => {
+    function handleClickOutside(event: MouseEvent) {
+      if (ref.current && !ref.current.contains(event.target as Node)) {
+        onClose();
+      }
+    }
+
+    document.addEventListener('mousedown', handleClickOutside);
+
+    return () => {
+      document.removeEventListener('mousedown', handleClickOutside);
+    };
+  }, [ref, onClose]);
+}
